third_party/blink/web_tests/platform/win11-arm64/http/tests/security/contentSecurityPolicy/sandbox-in-http-header-expected.txt

CONSOLE MESSAGE: Script executed in iframe.
This is a testharness.js-based test.
FAIL contentDocument should return null for sandboxed iframe assert_equals: expected null but got Document node with 1 child
Harness: the test ran to completion.
